requirements.txt
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==46.0.7
et_xmlfile==2.0.0
iniconfig==2.3.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.1
pandas==3.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.20.0
pypdfium2==5.7.0
pytest==9.0.3
python-dateutil==2.9.0.post0
RapidFuzz==3.14.5
six==1.17.0
tzdata==2026.1
